fix/relocate-periodic-timer
